requirements.txt
streamlit>=1.38
akshare>=1.13
pandas>=2.2
numpy>=1.26
plotly>=5.22
futu-api>=6.9

